tools/intel_dp_compliance: Add missing GLIB_CFLAGS
intel_dp_compliance.h includes the glib.h header file but the Makefile
does not explicitly pass a -I option with the path containing that
header, hence causing the build to fail. Note that this doesn't seem to
happen with a recent enough version of cairo, which implicitly provides
the correct -I option.

Signed-off-by: Thierry Reding <treding@nvidia.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>