tests/perf_pmu: PMU enable race test
Test that the PMU can be safely enabled in face of interrupt-heavy load on
an engine.

The test is probabilistic so run it for ten seconds in a loop to increase
the odds of hitting the race.

v2: Repeat the test a few times, until a timeout. (Chris Wilson)
v3: Added note in code and commit about probabilistic nature of the test.
    (Chris Wilson)

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>