meson: Don't build docs when cross-compiling
Building documentation requires executing all test binaries to produce
their respective description texts. This can be a very time-consuming
process, if viable at all, when the target arch differs from the
host.

Don't process the doc directory at all when an exe wrapper is
setup. This avoids the runtime penalty when the target binaries are
executed through qemu, but leaves everything as-is when the target
binaries can be directly executed, like when cross-compiling to x86
from x86_64 host.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Eric Anholt <eric@anholt.net>
Reviewed-by: Eric Anholt <eric@anholt.net>