igt/gem_fd_exhaustion: Actually test error handling on fd exhaustion
Long time ago, there was an issue with driver crashing due to lack of
error handling on struct file creation (gem_create()). We're unable to hit
this scenario, because dup() does not create filp. Let's replace dup() with
open() and assert that error is returned on exhaustion.

v2: Pass if gem_create fails on exhaustion, no matter the errno.

Signed-off-by: Karol Krol <karol.krol@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>