igt: Add gem_ctx_freq to exercise requesting freq via sysfs
Once proposed long ago was per-context frequency controls. As part of
that this test suite was written, which among its tests covered the
interaction of sysfs with the per-context controls. While we wait again
for approval of the per-context API, we can at least leverage the sysfs
tests to provide sorely lacking coverage of the sysfs knobs.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>