tests/kms_pipe_crc_basic: Remove legacy crc tests
Commit 8c6b709d96cb9 ("drm/i915: Use new CRC debugfs API") has been in
the kernel for more than a year.
In commit 6d16875736b9f ("lib/igt_debugfs: Remove support for legacy CRC api.")
we've removed support for the legacy CRC testing and we're about to remove
the last remnants from the kernel, so it's time to remove the last tests for
legacy CRC as well.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>