tests/pm_rpm: Don't try to create an X-tiled ARGB8888 framebuffer
On most platforms ARGB8888 is only supported by the cursor plane,
which doesn't support tiled buffers. Hence the kernel is under no
obligation to accept the request to create a tiled ARGB888
framebuffer. Apparently we've been letting this slip through so
far, but that is about to change.

Fortunately the resulting fb was neer actually used (in fact the
kernel would have rejected the setplane/atomic ioctl with such
a framebuffer). All we're really interested here is the gem BO
which we feed to the kernel via the setcursor ioctl after changing
the tiling to linear. To avoid changing the test drastically we
can simply change the fb to XRGB8888, which combined with X-tiling
is supported on all platforms.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>