tests/kms_plane_scaling: Change minimum fb height for YUV 420 planar
When the plane scaling is used with YUV 420 planar formats,
the height should be a minimum of 16 scanlines as per BSpec.
Minimum of 8 scanlines is for non-YUV 420 planar formats.
Patch changes the minimum fb height being used in the test
in case of YUV 420 planar (currently done for NV12, as
that is the only planar format supported)

v2: Addressed review comments from Maarten

Signed-off-by: Uma Shankar <uma.shankar@intel.com>
Signed-off-by: Vidya Srinivas <vidya.srinivas@intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>