tests/kms_frontbuffer_tracking: Fix Y alignment in the scaledprimary subtest, v3.
This is hidden behind the other kms_frontbuffer_tracking failures,
but is its own fail.

On gen9/gen10 we fail the scaledprimary tests because FBC refuses
to be enabled with "plane Y offset is misaligned".
Looking at the kernel, this is a workaround for FIFO underruns
which can be tested as well. Test that with the right alignment, FBC
is enabled and with misalignment FBC is disabled on <= gen10.

Changes since v1:
- Always check both alignments.
Changes since v2:
- Fix gen assignment.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=105678
Reviewed-by: Marta Lofstedt <marta.lofstedt@intel.com>