tests/gem_eio: Speed up test execution
If we stop relying on regular GPU hangs to be detected, but trigger them
manually as soon as we know our batch of interest is actually executing
on the GPU, we can dramatically speed up various subtests.

This is enabled by the pollable spin batch added in the previous patch.

v2:
 * Test gem_wait after reset/wedge and with reset/wedge after a few
   predefined intervals since gem_wait invocation. (Chris Wilson)

v3:
 Chris Wilson:
 * Decrease short test to 1us.
 * Use POSIX timers instead of signals to avoid interrupting gem_wait.
 * Improve comment.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Suggested-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Antonio Argenziano <antonio.argenziano@intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>