tests/kms_color: Do not mask LUT for non i915 drivers.
Masking assumes a direct relationship between the software LUT
structure, and hardware LUT. This is not always the case.

On AMD hardware for example, the hardware LUT is composed of
piecewise-linear segments, with end-point spaced exponentially along the
X axis, while software LUT is spaced linearly. Masking the LUT for the
purpose of truncating the resulting colors won't work here.

v2: Add commit message and sign off.

Signed-off-by: Leo (Sunpeng) Li <sunpeng.li@amd.com>
Acked-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>