gtkdoc: Remove the build_by_default hack
meson 0.45 properly supports generated content_files. Except for the
minor issue that we're hitting an internal bug in meson. I've made a
pull request for that

https://github.com/mesonbuild/meson/pull/3189

which has now been addressed in

commit c1f275bfa644beafab9f8572351d4b64d61c148b
Author: Nirbheek Chauhan <nirbheek@centricular.com>
Date:   Sun May 6 20:09:49 2018 +0530

    gnome.gtkdoc: Allow passing file objects as xml_files

    If we pass a source files() object, we will look for it in the build
    directory, which is wrong. If we pass a build files() object (from
    configure_file()), we will find it in the build directory, and then
    try to copy it on top of itself in gtkdochelper.py getting a
    SameFileError.

    Add a test for it, and also properly iterate custom target outputs
    when adding to content files.

Downside of this all is that we'd need to up the meson requirements to
something like 0.47, which isn't even released yet :-/

v2: Rebase and also add version requirement.

v3: I figured out how to make this work with a meson version check!

v4: Remove stray hunk (Petri).

Cc: Eric Anholt <eric@anholt.net>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>